requirements.txt
streamlit
sentence-transformers
pdfplumber
python-docx
PyPDF2
scikit-learn
numpy
torch
